Move images to separate module
