+++ src/app.module.ts
@@ -12,12 +12,11 @@ import { ProductsModule } from "./products/products.module";
 import { User } from "./users/entities/user.entity";
 import { Product } from "./products/entities/product.entity";
 import { Comment } from "./comments/comment.entity";
-import { Image } from "./products/entities/image.entity";
+import { Image } from "./images/image.entity";
 import { Rating } from "./products/entities/rating.entity";
 import { Order } from "./orders/entities/order.entity";
 import { OrdersModule } from "./orders/orders.module";
 import { OrderRow } from "./orders/entities/order-row.entity";
-import { CategoriesModule } from "./categories/categories.module";
 import { Category } from "./categories/category.entity";
 
 @Module({
