+++ src/images/dto/create-image.dto.ts
