+++ src/images/dto/update-image.dto.ts
