+++ src/images/images.service.ts
@@ -1,9 +1,9 @@
 import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
-import { Image } from "../entities/image.entity";
-import { CreateImageDto } from "../dto/create-image.dto";
-import { UpdateImageDto } from "../dto/update-image.dto";
+import { Image } from "./image.entity";
+import { CreateImageDto } from "./dto/create-image.dto";
+import { UpdateImageDto } from "./dto/update-image.dto";
 
 
 @Injectable()
