+++ src/products/services/products.service.ts
@@ -5,7 +5,7 @@ import { Product } from "../entities/product.entity";
 import { CreateProductDto } from "../dto/create-product.dto";
 import { UpdateProductDto } from "../dto/update-product.dto";
 import { Comment } from "../../comments/comment.entity";
-import { Image } from "../entities/image.entity";
+import { Image } from "../../images/image.entity";
 import { OrderRow } from "../../orders/entities/order-row.entity";
